telegram_bot/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
aiosqlite==0.19.0
